test: organize and stabilize e2e acceptance suite
